Reorganized directory to be more standard
