+++ xray/__init__.py
@@ -1,7 +1,7 @@
 from .xarray import as_xarray, XArray, CoordXArray, broadcast_xarrays
 from .conventions import decode_cf_datetime, encode_cf_datetime
 from .dataset import Dataset, open_dataset
-from .dataset_array import DataArray, align
+from .data_array import DataArray, align
 from .utils import orthogonal_indexer, xarray_equal
 
 from .version import version as __version__
