+++ xray/common.py
